gpio: Add mraa_get_gpiochip_base
This allows to retrieve the GPIO base using a stable device path as
found in sysfs. We cannot rely on the absolute GPIO numbers because
Linux prefers dynamic numbering. So, different probe ordering or
additional GPIO chips can ruin our day if we just use hard-coded
numbers.

This is a building block to enable upstream kernels on Galileo Gen2
and IOT2000.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>