intel_galileo_rev_g.c: Provide gpio_mode_replace stub for upstream kernels
To set pull-up/pull-down for output pins, mraa so far relies on vendor
kernel extensions that are unlikely to show up in this form in upstream.
Fortunately, we do not need this feature in most scenarios, e.g. the
enabling of I2C on the Galileo Gen2 and the IOT2000.

This detect the kernel we are running on - indirectly via the
availability of the PWM1 ACPI path - and installs a gpio_mode_replace
stub in case of upstream. We can accept mode switching to HIZ because
this is what the hardware does implicitly for input pins. All other
requests are rejected as unimplemented.

Tested via mraa-i2c on an IOT2040, both with yocto and upstream kernel.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>